Move prefs package to commons
