+++ kirinmaru-android/src/main/java/commons/android/core/prefs/PrefDynamicModel.kt
@@ -1,5 +1,6 @@
-package stream.reconfig.kirinmaru.android.util.prefs
+package commons.android.core.prefs
 
+import android.annotation.SuppressLint
 import android.content.SharedPreferences
 import android.support.annotation.AnyThread
 import android.support.annotation.WorkerThread
@@ -10,10 +11,10 @@ import android.support.annotation.WorkerThread
  * [T] can be any valid SharedPreferences primitive type (String set included)
  */
 abstract class PrefDynamicModel<T, in P>(
-    private val default: T,
-    private inline val retrieve: SharedPreferences.(key: String, default: T) -> T,
-    private inline val store: SharedPreferences.Editor.(key: String, data: T) -> Unit,
-    private val prefs: SharedPreferences
+  private val default: T,
+  private inline val retrieve: SharedPreferences.(key: String, default: T) -> T,
+  private inline val store: SharedPreferences.Editor.(key: String, data: T) -> Unit,
+  private val prefs: SharedPreferences
 ) {
 
   abstract fun key(input: P): String
@@ -21,8 +22,10 @@ abstract class PrefDynamicModel<T, in P>(
   @WorkerThread
   open fun load(keyedBy: P, defaultVal: T = default): T = retrieve(prefs, key(keyedBy), defaultVal)
 
+  @SuppressLint("ApplySharedPref")
   @AnyThread
-  open fun persist(keyedBy: P, data: T) {
-    prefs.edit().also { store(it, key(keyedBy), data) }.apply()
+  open fun persist(keyedBy: P, data: T, commit: Boolean = false) {
+    prefs.edit().also { store(it, key(keyedBy), data) }
+      .run { if (commit) commit() else apply() }
   }
 }
